+++ CHANGELOG.rst
@@ -15,6 +15,11 @@ entries are sorted in descending chronological order.
 Unreleased
 ==========
 
+Added
+-----
+- `EBSD.static_background` property for easier access to the background pattern.
+  (`#475 <https://github.com/pyxem/kikuchipy/pull/475>`_)
+
 Fixed
 -----
 - Hopefully prevent EBSD refinement tests using random data to fail on Azure.